exceptions added

+++ clinicaOdontologica/src/main/java/com/dh/clinicaOdontologica/services/OdontologoService.java
@@ -1,6 +1,7 @@
 package com.dh.clinicaOdontologica.services;
 
 import com.dh.clinicaOdontologica.dto.OdontologoDTO;
+import com.dh.clinicaOdontologica.exceptions.NotFoundException;
 import com.dh.clinicaOdontologica.models.Odontologo;
 import com.dh.clinicaOdontologica.repositories.OdontologoRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -8,7 +9,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -29,11 +29,15 @@ public class OdontologoService {
     public Odontologo actualizarOdontologo(Odontologo odontologo){
         return guardarOdontologo(odontologo);
     }
-    public void eliminarOdontologo(Long id){
+    public void eliminarOdontologo(Long id) throws NotFoundException {
+        if (odontologoRepository.findById(id).isEmpty())
+            throw new NotFoundException("el odontologo con id: " + id.toString() + " no existe");
         odontologoRepository.deleteById(id);
     }
 
-    public OdontologoDTO buscarOdontologo(Long id){
+    public OdontologoDTO buscarOdontologo(Long id) throws NotFoundException {
+        if (odontologoRepository.findById(id).isEmpty())
+            throw new NotFoundException("el odontologo con id: " + id.toString() + " no existe");
         return mapper.convertValue(odontologoRepository.findById(id),OdontologoDTO.class);
 
     }
